fix: wasm64 runtime fixes — WASI iovec, dylink table64, LP64 structs
- WASI shim: update iovec reads to 16-byte stride (i64 base + i64 len)
  for fd_pread, fd_pwrite, sock_recv, sock_send
- WASI fixtures: rebuild wasi-hello.wat and wasi-args.wat for memory64
  iovec layout (i64.store, 16-byte entries)
- dylink.ts: use i64 globals for __memory_base/__table_base, handle
  table64 (BigInt) vs table32 for module-exported tables, fix GOT
  entries and export relocation for wasm64
- dylink.test.ts: create table64 via wasm module export (V8's JS API
  new WebAssembly.Table({index:"i64"}) creates i32 table internally)
- dlopen-e2e.test.ts: add -mwasm64 to wasm-ld, --allow-multiple-definition
- dlopen example build.sh: update for wasm64 target
- kernel IPC: fix mtype read/write for LP64 (i64 msgbuf.mtype)
- kernel wasm_api: fix itimerval/itimerspec struct sizes for LP64
- worker-main.ts: fix TLS/stack pointer handling for i64 globals
- centralized-test-helper.ts: memory64 params for test helper
- Test fixes for memory64, table64, BigInt across all test files
- Rebuild wasm binaries (kernel, exec-caller/child, fork-exec, mmap_shared_test)
- Add musl-overlay/src/thread/wasm64posix/ for LP64 thread support

+++ crates/kernel/src/ipc.rs
@@ -44,7 +44,7 @@ const SEMMSL: usize = 32;       // max semaphores per set
 
 /// A single message in a SysV message queue.
 struct MsgEntry {
-    mtype: i32,
+    mtype: i64,
     data: Vec<u8>,
 }
 
@@ -91,7 +91,7 @@ pub struct MsgQueueInfo {
 /// Result of msgrcv.
 #[derive(Debug)]
 pub struct MsgRcvResult {
-    pub mtype: i32,
+    pub mtype: i64,
     pub data: Vec<u8>,
 }
 
@@ -274,7 +274,7 @@ impl IpcTable {
     }
 
     /// Send a message to a queue.
-    pub fn msgsnd(&mut self, qid: i32, mtype: i32, data: &[u8], flags: u32, pid: u32) -> Result<(), Errno> {
+    pub fn msgsnd(&mut self, qid: i32, mtype: i64, data: &[u8], flags: u32, pid: u32) -> Result<(), Errno> {
         if mtype <= 0 {
             return Err(Errno::EINVAL);
         }
@@ -309,7 +309,7 @@ impl IpcTable {
         &mut self,
         qid: i32,
         max_size: u32,
-        msgtype: i32,
+        msgtype: i64,
         flags: u32,
         pid: u32,
     ) -> Result<MsgRcvResult, Errno> {
